move monitoring init from Monitoring.cpp to init.h
