+++ model/CMakeLists.txt
@@ -82,6 +82,7 @@ set (Model_CPP
   Host/WithinHost/Treatments.cpp
   
   mon/Monitoring.cpp
+  mon/init.cpp
   mon/Continuous.cpp
   
   util/errors.cpp
