+++ model/main.cpp
@@ -31,6 +31,7 @@
 #include "util/XMLChecker.h"
 
 #include "mon/Continuous.h"
+#include "mon/init.h"
 #include "mon/Monitoring.h"
 
 #include "interventions/InterventionManager.h"
